Move the item cache entry type next to the item cache
(and remove outdated comment that caused doc error)
